+++ PlannerApi/Deserializers/PriceDeserializer.cs
@@ -4,8 +4,8 @@ using Microsoft.Azure.KeyVault.Models;
 namespace events_planner.Deserializers {
     public class PriceDeserializer {
         
-        [Range(0, int.MaxValue, ErrorMessage = "Price cannot be under 0 value")]
-        public int Amount { get; set; }
+        [Range(0, float.MaxValue, ErrorMessage = "Price cannot be under 0 value")]
+        public float Amount { get; set; }
         
         public int EventId { get; set; }
